Use GC_ERROR macro for reporting all errors
Since gc is a library, we can't decide what to do when a runtime error occurs.
Therefore, we report errors using the GC_ERROR macro, which the user can
define according to his needs before including a gc header.  If this macro is
not defined when including a gc header, nothing happens when an error occurs.

This obsoletes the GC_MALLOC_FAILED macro.